refactor: refactor price-blocks for mini-cart
