+++ src/UI/pages/common-models/states-models.ts
@@ -4,7 +4,6 @@ import {
   localBasketItemInit,
   SymbolCurrency,
 } from './common-models';
-import MainCart from '../product-cards/main-cart/main-cart';
 
 export interface IMainPageState {
   error: null;
@@ -100,4 +99,8 @@ export interface ICartCountProviderState {
 
 export interface INavigateBlockState {
   currentCategory: string;
+}
+export interface IPriceBlockState {
+  symbol: string;
+  amount: number;
 }
