+++ src/UI/pages/product-cards/minicart-pages/mini-cart-basic-block/mini-cart-basic-block.tsx
@@ -1,11 +1,8 @@
 import React from 'react';
 import styles from './mini-cart-basic-block.module.css';
-import MainCartBasicBlock from '../../main-cart/main-cart-blocks/main-cart-basic-block/main-cart-basic-block';
 import { IProductAttribute } from '../../../common-models';
 import MiniCartCommonAttributes from '../mini-cart-common-attributes/mini-cart-common-attributes';
 import MiniPriceBlock from '../mini-cart-blocks/blocks/mini-price-block/mini-price-block';
-import PriceBlock from '../../../common/common-bloks/price-block/price-block';
-import PriceBlockAbstractClass from '../../main-cart/abstract-classes/PriceBlockAbstractClass';
 import BasicBlock from '../../pdp-card/pdp-card-blocks/basic-block';
 
 class MiniCartBasicBlock extends BasicBlock {
@@ -13,7 +10,7 @@ class MiniCartBasicBlock extends BasicBlock {
     const arrAttributes = this.props.product.attributes as IProductAttribute[];
     const productId =
       this.props.product.id === '' ? 'xbox-series-s' : this.props.product.id;
-    const symbolCurrency = this.props.currentCurrency;
+    const prices = Array.isArray(this.props.product.prices) ? this.props.product.prices : [this.props.product.prices];
     const isEmpty: boolean = this.props.product.id === '';
     const brand = isEmpty ? 'Your cart is empty' : this.props.product.brand;
 
@@ -29,9 +26,7 @@ class MiniCartBasicBlock extends BasicBlock {
         <div className={styles.priceBlock}>
           <div>
             <MiniPriceBlock
-              id={productId}
-              symbolCurrency={symbolCurrency}
-              isEmpty={isEmpty}
+              prices={prices}
             />
           </div>
         </div>
