+++ src/UI/pages/product-cards/minicart-pages/mini-cart.tsx
@@ -2,10 +2,9 @@ import React from 'react';
 import stylesCart from './mini-cart.module.css';
 import MiniCartItem from './mini-cart-blocks/mini-cart-item';
 import MiniButtonBlock from './mini-cart-blocks/mini-button-block';
-// import MainCart from '../main-cart/main-cart';
 import {
   ICurrency,
-  IMainCartProps, IMainCartState,
+  IMainCartProps,
   localBasketItemInit,
   SymbolCurrency,
 } from '../../common-models';
@@ -13,8 +12,7 @@ import MiniTotalBlock from './mini-total-block/mini-total-block';
 import { LOCAL_BASKET } from '../../../../constants';
 import { State } from '../../../../store/store';
 import { renewBasket } from '../../../../store/cartSlice';
-import { connect, DispatchProp } from 'react-redux';
-// import { AnyAction } from '@reduxjs/toolkit';
+import { connect } from 'react-redux';
 import CartAbstractClass from '../main-cart/abstract-classes/CartAbstractClass';
 
 const mapStateToProps = (state: State) => {
@@ -23,9 +21,7 @@ const mapStateToProps = (state: State) => {
 const mapDispatchToPropsFactory = { renewBasket };
 
 type IProps = Readonly<IMainCartProps>;
-type IState = Readonly<IMainCartState>;
 
-// class MiniCart extends connect(mapStateToProps, mapDispatchToPropsFactory)(MainCart) {
 class MiniCart extends CartAbstractClass {
   constructor(props: IProps) {
     super(props);
